Method of Content-Type negotiation. Am unsure about the way it works -
I could have the shell expansion syntax provided by path, but it would add a bit of un-needed wasted
CPU usage, when the only side effect is an edge-case in which you cannot have a.png and a.xcf in the same
IgnoreExtensions router.

I could solve this problem by having a special Router that can demux this route in this case, or, alternately
-- and I think somewhat more semantically, I can have a PathRouter that is passed to a function and takes note of
a wildcard-based grammar to build several maps, each with a common prefix/suffix, and check those in turn.

This has the advantage of forming a near-identical process after construction to a vanilla PathRouter,
since one map would be constructed with an empty prefix.

It also nicely prevents the user having to make special handler functions if they want to do special
in-directory parameters like '/files/yy,mm,dd' since we can specify something like:

PathRouter {
	files: PathGrammar(
		*,*,*, fw.HandlerOf(myHandler),
   ),
}

One of the advantages of framework eight is that everything save the raw strings that are matched
in the routing is pure Go code, with no special routing grammar or regex, which improves both
speed and semantics.

In the case of the comma-separated dates, it might be best that the API user writes their own
PathRouter, which fits into the path routing tree. Perhaps improvement of the PathRouter API is the
solution to this problem?